refactor: extract environment types to dedicated module
Move Config, Handles, Env, and ServerConfig from Hoard.Effects
to new Hoard.Types.Environment module to better separate concerns
between configuration/runtime data and effect stack implementation.

Update imports in Server, Config.Loader, and TestHelpers to use
the new module directly.
