+++ src/Hoard/Server.hs
@@ -8,7 +8,8 @@ import Network.Wai.Handler.Warp (defaultSettings, runSettings, setHost, setPort)
 import Servant
 
 import Hoard.API (API, server)
-import Hoard.Effects (AppEff, Config (..), Env (..), ServerConfig (..), runEffectStack)
+import Hoard.Effects (AppEff, runEffectStack)
+import Hoard.Types.Environment (Config (..), Env (..), ServerConfig (..))
 
 import Hoard.Effects.Conc qualified as Conc
 import Hoard.Effects.Log qualified as Log
@@ -18,13 +19,19 @@ runServer :: (AppEff es) => Env -> Eff es ()
 runServer env = do
     _ <- Conc.fork $ do
         -- Log startup messages
-        let ServerConfig {host = serverHost, port = serverPort} = env.config.server
-            host = toString serverHost
-            port = fromIntegral serverPort
+        let host = toString env.config.server.host
+            port = fromIntegral env.config.server.port
+
         Log.debug $ "Starting Hoard server on " <> toText host <> ":" <> show port
 
         -- Run Warp server (needs liftIO since Warp's runSettings is in IO)
         let settings = setPort port $ setHost (fromString host) defaultSettings
-            servantApp = hoistServer (Proxy @API) (Handler . runEffectStack env) Hoard.API.server
+            servantApp =
+                hoistServer
+                    (Proxy @API)
+                    (Handler . runEffectStack env)
+                    Hoard.API.server
+
         liftIO $ runSettings settings (serve (Proxy @API) servantApp)
+
     pure ()
